added misc action and reducer. created the main page component to render business list below tabs.
