+++ Actions/miscs.js
@@ -1,0 +1,7 @@
+export const CREATE_MISC = 'create_misc';
+
+export function createMisc() {
+  return {
+    type: CREATE_MISC,
+  };
+};
