+++ Components/Header.js
@@ -1,9 +1,7 @@
 import React, { Component } from 'react';
 import { Text, View, TextInput, StyleSheet, Dimensions } from 'react-native';
 import { connect } from 'react-redux';
-import shuffle from 'lodash.shuffle';
 
-import BusinessList from './BusinessList';
 import { setSearchValue } from '../Actions/searchBar';
 import {
   fetchRestaurantsRequest,
@@ -31,27 +29,7 @@ class Header extends Component {
     this.props.fetchLaundromatsRequest(`https://owl-oiche-yelp-api.herokuapp.com/api/yelpResults${makeQuery({ location: this.props.searchBar, term: 'laundry' })}`);
   }
 
-  decideBusinessesToDisplay() {
-    if (this.props.activeTab === 'Restaurants') {
-      return this.props.businesses.restaurants;
-    }
-
-    if (this.props.activeTab === 'Pharmacies') {
-      return this.props.businesses.pharmacies;
-    }
-
-    if (this.props.activeTab === 'Wifi') {
-      return this.props.businesses.wifiSpots;
-    }
-
-    else {
-      let misc = this.props.businesses.misc;
-      return shuffle(misc.gasStations.concat(misc.groceries, misc.laundromats));
-    }
-  }
-
   render() {
-    const businesses = this.decideBusinessesToDisplay();
     return (
       <View>
         <View style={styles.topBuffer}>
@@ -67,18 +45,13 @@ class Header extends Component {
                      onSubmitEditing={(submit) => this.searchSubmit(submit)}
           />
         </View>
-        { this.props.businesses.restaurants.length > 0 ? (
-          <BusinessList businesses={businesses} />
-        ) : (
-          <Text>We need to know your location, please enter a city above.</Text>
-        )}
       </View>
     );
   }
 }
 
-function mapStateToProps({ activeTab, searchBar, businesses }) {
-  return { activeTab, searchBar, businesses };
+function mapStateToProps({ activeTab, searchBar }) {
+  return { activeTab, searchBar };
 }
 
 const actions = {
